them ngan sach và giao dich

+++ QLCT/src/main/java/com/myapp/QLCT/repository/TransactionRepository.java
@@ -11,7 +11,7 @@ import com.myapp.QLCT.entity.Transaction;
 import java.util.List;
 
 @Repository
-public interface TransactionRepository extends JpaRepository<Transaction, Long> {
+public interface TransactionRepository extends JpaRepository<Transaction, Integer> {
 
     @Query("SELECT COALESCE(SUM(t.amount), 0) FROM Transaction t " +
            "WHERE t.type = 'thu' " +
